requirements.txt
Flask
flask_cors
werkzeug
PyJWT
